\name{mmgrl}
\alias{mmgrl}
\title{An Example GRangesList Object}
\description{
A toy GRangesList dataset derived from the GRanges dataset in this package for 
purposes of demonstration.
}
\details{
The GRanges dataset was submitted to display(), subsetted and several 
iterations of the results were fused into a GRangesList object.  This is 
fake data.
}
\examples{
data(mmgrl)
}
\keyword{datasets}